[DGS-359] remove debug code
