+++ src/Controller/AbstractAdminController.php
@@ -30,6 +30,7 @@ use Invertus\dpdBaltics\OnBoard\Service\OnBoardStepActionService;
 use Invertus\dpdBaltics\Provider\CurrentCountryProvider;
 use Invertus\dpdBaltics\Provider\ImportExportURLProvider;
 use Invertus\dpdBaltics\Repository\ParcelShopRepository;
+use Invertus\dpdBaltics\Validate\Version\ModuleLatestVersionValidator;
 use Media;
 use ModuleAdminController;
 use Tools;
@@ -61,6 +62,7 @@ class AbstractAdminController extends ModuleAdminController
         }
 
         $this->displayTestModeWarning();
+        $this->checkCanUpdateModule();
         $this->initFlashMessages();
 
         parent::initContent();
@@ -246,6 +248,30 @@ class AbstractAdminController extends ModuleAdminController
             }
 
             unset($this->context->cookie->{$type});
+        }
+    }
+
+    private function checkCanUpdateModule()
+    {
+        /** @var ModuleLatestVersionValidator $moduleVersionValidator */
+        $moduleVersionValidator = $this->module->getModuleContainer()->get('invertus.dpdbaltics.validator.module_latest_version_validator');
+
+        try {
+            $isModuleVersionLatest = $moduleVersionValidator->validate();
+        } catch (\Exception $e) {
+            $this->errors[] = $e->getMessage();
+        }
+
+        if (!$isModuleVersionLatest) {
+            $this->warnings[] = $this->context->smarty->fetch(
+                $this->module->getLocalPath() . 'views/templates/admin/warning-message-with-link.tpl',
+                [
+                    'messageLink' => Config::DPD_GITHUB_REPO_RELEASE_LATEST_DOWNLOAD_URL,
+                    'messageStart' => $this->module->l('You are using an outdated version of the module. Please update the module to the latest version. ', self::FILENAME),
+                    'messageEnd' => '',
+                    'linkText' => $this->module->l('Download latest module', self::FILENAME)
+                ]
+            );
         }
     }
 }
